+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Changed
+- Rename Decrypt to DecryptAsync
+- Rename Encrypt to EncryptAsync
+- Using Awaitable functions inside ICryptographer
 - Increase Unity minimum version to 6000.1
 
 ## [1.0.0] - 2022-08-13